ipa: simple: awb: Port to use libipa AwbAlgorithm
Port the SoftISP Awb algorithm to use the new libipa implementation
of AwbAlgorithm.

The awbAlgo_ class member is initialized with the Q<2, 8> type even if
there is no physical register representation for SoftISP.

The usage of libipa::awb::Context, which defines the gains vector as
RGB<double>, in the SimpleIPA ActiveState and FrameContext requires
changes to debayer_cpu.cpp and blc.cpp in order not to mix RGB<float>
and RGB<double>.

Tested-by: Robert Mader <robert.mader@collabora.com>
Tested-by: Milan Zamazal <mzamazal@redhat.com>

+++ include/libcamera/internal/software_isp/debayer_params.h
@@ -21,10 +21,10 @@ struct DebayerParams {
 	Matrix<float, 3, 3> combinedMatrix = { { 1.0, 0.0, 0.0,
 						 0.0, 1.0, 0.0,
 						 0.0, 0.0, 1.0 } };
-	RGB<float> blackLevel = RGB<float>({ 0.0, 0.0, 0.0 });
+	RGB<double> blackLevel = RGB<double>({ 0.0, 0.0, 0.0 });
 	float gamma = 1.0;
 	float contrastExp = 1.0;
-	RGB<float> gains = RGB<float>({ 1.0, 1.0, 1.0 });
+	RGB<double> gains = RGB<double>({ 1.0, 1.0, 1.0 });
 };
 
 } /* namespace libcamera */
